linting

+++ contracts/interfaces/IWETH9.sol
@@ -4,19 +4,19 @@ pragma solidity ^0.7.6;
 interface IWETH9 {
     function deposit() external payable;
 
-    function withdraw(uint wad) external;
+    function withdraw(uint256 wad) external;
 
-    function totalSupply() external view returns (uint);
+    function totalSupply() external view returns (uint256);
 
-    function approve(address guy, uint wad) external returns (bool);
+    function approve(address guy, uint256 wad) external returns (bool);
 
-    function transfer(address dst, uint wad) external returns (bool);
+    function transfer(address dst, uint256 wad) external returns (bool);
 
     function transferFrom(
         address src,
         address dst,
-        uint wad
+        uint256 wad
     ) external returns (bool);
 
-    function balanceOf(address) external view returns (uint);
+    function balanceOf(address) external view returns (uint256);
 }
